+++ src/main/java/com/overclock/overclock/service/AssemblyService.java
@@ -2,6 +2,7 @@ package com.overclock.overclock.service;
 
 import com.overclock.overclock.model.Assembly;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
@@ -14,7 +15,7 @@ public interface AssemblyService {
     List<Assembly> getAllAssemblies();
     boolean save(Assembly assembly);
     boolean delete(BigInteger id);
-    boolean updateScore(BigInteger id, BigInteger newScore);
+    boolean updateScore(BigInteger id, BigDecimal newScore);
     void checkAssemblyValidation(Assembly assembly);
 }
